--- .schemathesis/default/cache/crashes/POST_v1_feedback_records_negative_data_rejection_4f5af2f0.json
@@ -1,166 +1,0 @@
-{
-  "operation": "POST /v1/feedback-records",
-  "method": "POST",
-  "path_template": "/v1/feedback-records",
-  "fingerprint": "4f5af2f0",
-  "case_id": "29Vpla",
-  "code_sample": "curl -X POST -H 'Authorization: [Filtered]' -H 'Content-Type: application/json' -d '{\"collected_at\": \"2000-01-01T00:00:00Z\", \"field_group_id\": \"feature_priority\", \"field_group_label\": \"Rank these features by importance\", \"field_id\": \"q1\", \"field_label\": \"Reports\", \"field_type\": \"nps\", \"language\": \"en\", \"metadata\": {\"question_type\": \"matrix\"}, \"source_id\": \"survey-123\", \"source_name\": \"NPS Survey\", \"source_type\": \"survey\", \"submission_id\": \"550e8400-e29b-41d4-a716-446655440000\", \"tenant_id\": \"org-123\", \"user_id\": \"user-abc-123\", \"value_boolean\": true, \"value_date\": \"2000-01-01T00:00:00Z\", \"value_id\": \"opt_very_satisfied\", \"value_number\": 1000000000000001, \"value_text\": \"Great service!\"}' --insecure http://localhost:8888/v1/feedback-records",
-  "sequence": [
-    {
-      "method": "POST",
-      "url": "http://localhost:8888/v1/feedback-records",
-      "url_template": "http://localhost:8888/v1/feedback-records",
-      "headers": {
-        "User-Agent": "schemathesis/4.24.3",
-        "Accept-Encoding": "gzip, deflate",
-        "Accept": "*/*",
-        "Connection": "keep-alive",
-        "Authorization": "[Filtered]",
-        "X-Schemathesis-TestCaseId": "29Vpla",
-        "Content-Type": "application/json",
-        "Content-Length": "610"
-      },
-      "response": {
-        "status_code": 201,
-        "headers": {
-          "content-type": "application/json",
-          "x-request-id": "019fd93e-2a93-7ae5-a8a4-6261e0ba24c6",
-          "date": "Thu, 06 Aug 2026 22:42:37 GMT",
-          "content-length": "703"
-        },
-        "body": "{\"id\":\"019fd93e-2aae-70fc-a0c2-a3392ce6e8e2\",\"collected_at\":\"2000-01-01T00:00:00Z\",\"created_at\":\"2026-08-06T22:42:37.869854Z\",\"updated_at\":\"2026-08-06T22:42:37.869854Z\",\"source_type\":\"survey\",\"source_id\":\"survey-123\",\"source_name\":\"NPS Survey\",\"field_id\":\"q1\",\"field_label\":\"Reports\",\"field_type\":\"nps\",\"field_group_id\":\"feature_priority\",\"field_group_label\":\"Rank these features by importance\",\"value_text\":\"Great service!\",\"value_id\":\"opt_very_satisfied\",\"value_number\":1000000000000001,\"value_boolean\":true,\"value_date\":\"2000-01-01T00:00:00Z\",\"metadata\":{\"question_type\":\"matrix\"},\"language\":\"en\",\"user_id\":\"user-abc-123\",\"tenant_id\":\"org-123\",\"submission_id\":\"550e8400-e29b-41d4-a716-446655440000\"}\n"
-      },
-      "link": null,
-      "checks": [
-        {
-          "name": "negative_data_rejection",
-          "status": "failure",
-          "message": "API accepted schema-violating request\n\nInvalid data should have been rejected\nExpected: 400, 401, 403, 404, 405, 406, 409, 422, 428, 5xx\nInvalid component: in body - value_number: Value greater than maximum"
-        }
-      ],
-      "meta": {
-        "generation": {
-          "time": 5.790963768959045e-06,
-          "mode": "negative"
-        },
-        "components": {
-          "HEADER": "positive",
-          "BODY": "negative"
-        },
-        "phase": {
-          "name": "coverage",
-          "data": {
-            "type": "CoveragePhaseData",
-            "description": "value_number: Value greater than maximum",
-            "location": "/properties/value_number/maximum",
-            "parameter": "application/json",
-            "parameter_location": "BODY",
-            "scenario": "value_above_maximum"
-          }
-        },
-        "pool_draws": [
-          {
-            "location": "body",
-            "parameter_name": "field_id",
-            "resource_name": "Field",
-            "resource_field": "field_id",
-            "source_operation": "POST /v1/feedback-records",
-            "source_status": 201
-          },
-          {
-            "location": "body",
-            "parameter_name": "field_group_id",
-            "resource_name": "FieldGroup",
-            "resource_field": "field_group_id",
-            "source_operation": "POST /v1/feedback-records",
-            "source_status": 201
-          },
-          {
-            "location": "body",
-            "parameter_name": "source_id",
-            "resource_name": "Source",
-            "resource_field": "id",
-            "source_operation": "POST /v1/feedback-records",
-            "source_status": 201
-          },
-          {
-            "location": "body",
-            "parameter_name": "submission_id",
-            "resource_name": "Submission",
-            "resource_field": "submission_id",
-            "source_operation": "POST /v1/feedback-records",
-            "source_status": 201
-          },
-          {
-            "location": "body",
-            "parameter_name": "tenant_id",
-            "resource_name": "Tenant",
-            "resource_field": "tenant_id",
-            "source_operation": "POST /v1/feedback-records",
-            "source_status": 201
-          },
-          {
-            "location": "body",
-            "parameter_name": "user_id",
-            "resource_name": "User",
-            "resource_field": "user_id",
-            "source_operation": "POST /v1/feedback-records",
-            "source_status": 201
-          },
-          {
-            "location": "body",
-            "parameter_name": "value_id",
-            "resource_name": "Value",
-            "resource_field": "value_id",
-            "source_operation": "POST /v1/feedback-records",
-            "source_status": 201
-          }
-        ],
-        "pool_misses": [],
-        "semantic_draws": [],
-        "dictionary_draws": [],
-        "constants_draws": [],
-        "raw_containers": {
-          "HEADER": {
-            "Authorization": "Bearer "
-          }
-        }
-      },
-      "path": "/v1/feedback-records",
-      "path_parameters": {},
-      "query": {},
-      "case_headers": {
-        "Authorization": "[Filtered]"
-      },
-      "cookies": {},
-      "case_body": {
-        "encoding": "json",
-        "value": {
-          "collected_at": "2000-01-01T00:00:00Z",
-          "field_group_id": "feature_priority",
-          "field_group_label": "Rank these features by importance",
-          "field_id": "q1",
-          "field_label": "Reports",
-          "field_type": "nps",
-          "language": "en",
-          "metadata": {
-            "question_type": "matrix"
-          },
-          "source_id": "survey-123",
-          "source_name": "NPS Survey",
-          "source_type": "survey",
-          "submission_id": "550e8400-e29b-41d4-a716-446655440000",
-          "tenant_id": "org-123",
-          "user_id": "user-abc-123",
-          "value_boolean": true,
-          "value_date": "2000-01-01T00:00:00Z",
-          "value_id": "opt_very_satisfied",
-          "value_number": 1000000000000001,
-          "value_text": "Great service!"
-        }
-      },
-      "media_type": "application/json",
-      "parent_index": null
-    }
-  ]
-}
